+++ __tests__/__configs__/jest.config.js
@@ -1,0 +1,6 @@
+const path = require('path');
+
+module.exports = {
+  rootDir: path.join(__dirname, '../'),
+  testPathIgnorePatterns: ['/node_modules/', '/__configs__/'],
+};